generated comparison data set
created a python file which generates a data set to which we can compare images. each number.png represents the lunar day of it's number.
the script generates a shadow based on a phase from -1 to 1 with 0 being the full moon ; -1 and 1 being the new moons. the shadow is then overlayed on the base image which is just a png of a white circle. this can of course be changed